extracted the shape creation into functions and these functions into seperate files
