+++ js-structure/js-structure_shapes/index.html
@@ -9,7 +9,7 @@
     <link rel="shortcut icon" href="./public/favicon.png" type="image/x-icon" />
     <title>JS Structure: Shapes</title>
     <link rel="stylesheet" href="./styles/styles.css" />
-    <script src="./index.js" defer></script>
+    <script src="./index.js" type="module" defer></script>
   </head>
   <body id="root"></body>
 </html>
