+++ js-structure/js-structure_shapes/utils/randomColor.js
@@ -1,3 +1,3 @@
-function getRandomColor() {
+export function getRandomColor() {
   return `hsl(${Math.floor(Math.random() * 360)} 80% 55%)`;
 }
